# contracts/openapi/music.yaml
openapi: 3.1.0
...
info:
  title: music-view API
  version: 1.0.0
  description: |
    Narzędzie DJ-a do zarządzania biblioteką muzyczną. Rozdział danych:
    `track_catalog` trzyma fakty o utworze, `library_entry` prywatne dane DJ-a.
servers:
  - url: /
    description: aplikacja za proxy /api
...
paths:
...
  /api/playlists:
    get:
      tags: [playlists]
      operationId: listPlaylists
      responses:
        '200':
          description: Lista playlist
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PlaylistSummaryResponse' }
    post:
      tags: [playlists]
      operationId: createPlaylist
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SavePlaylistRequest' }
      responses:
        '201':
          description: Playlista utworzona
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlaylistSummaryResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/playlists/{id}:
    get:
      tags: [playlists]
      operationId: getPlaylist
      parameters: [{ $ref: '#/components/parameters/PlaylistId' }]
      responses:
        '200':
          description: Playlista ze składem
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlaylistResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [playlists]
      operationId: renamePlaylist
      parameters: [{ $ref: '#/components/parameters/PlaylistId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SavePlaylistRequest' }
      responses:
        '200':
          description: Playlista po zmianie
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlaylistSummaryResponse' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [playlists]
      operationId: deletePlaylist
      parameters: [{ $ref: '#/components/parameters/PlaylistId' }]
      responses:
        '204': { description: Usunięto }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/playlists/{id}/tracks:
    post:
      tags: [playlists]
      operationId: addPlaylistTrack
      summary: Dokłada utwór na koniec setu
      parameters: [{ $ref: '#/components/parameters/PlaylistId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddPlaylistTrackRequest' }
      responses:
        '200':
          description: Playlista po zmianie
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlaylistResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    put:
      tags: [playlists]
      operationId: reorderPlaylist
      summary: Nowa kolejność — musi być permutacją obecnego składu
      parameters: [{ $ref: '#/components/parameters/PlaylistId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReorderPlaylistRequest' }
      responses:
        '200':
          description: Playlista po zmianie kolejności
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlaylistResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/playlists/{id}/tracks/{spotifyId}:
    delete:
      tags: [playlists]
      operationId: removePlaylistTrack
      summary: Usuwa utwór i zwiera pozycje pozostałych
      parameters:
        - { $ref: '#/components/parameters/PlaylistId' }
        - { $ref: '#/components/parameters/SpotifyId' }
      responses:
        '200':
          description: Playlista po zmianie
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlaylistResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/playlists/{id}/export-to-spotify:
    post:
      tags: [playlists]
      operationId: exportPlaylistToSpotify
      summary: Pierwszy eksport zakłada playlistę, kolejny nadpisuje tę samą
      parameters: [{ $ref: '#/components/parameters/PlaylistId' }]
      responses:
        '200':
          description: Wynik eksportu
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlaylistExportResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/sets/propose:
    post:
      tags: [sets]
      operationId: proposeSet
      summary: Propozycja setu z puli kandydatów; nic nie zapisuje
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetProposalRequest' }
      responses:
        '200':
          description: Propozycja
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SetProposalResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/sets/{playlistId}/fill:
    post:
      tags: [sets]
      operationId: fillSet
      summary: Uzupełnia gotowy set do zadanego czasu
      parameters: [{ $ref: '#/components/parameters/PlaylistIdPath' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetFillRequest' }
      responses:
        '200':
          description: Dalszy ciąg setu
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SetFillResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/sets/{playlistId}/suggest:
    post:
      tags: [sets]
      operationId: suggestForSet
      summary: Dobiera kandydatów na wskazane miejsce w secie
      parameters: [{ $ref: '#/components/parameters/PlaylistIdPath' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetSuggestionRequest' }
      responses:
        '200':
          description: Propozycje
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SetSuggestionResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
...
components:
  parameters:
    SpotifyId:
      name: spotifyId
      in: path
      required: true
      schema: { type: string }
    PlaylistId:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    PlaylistIdPath:
      name: playlistId
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  responses:
    BadRequest:
      description: Błąd walidacji
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Zasób nie istnieje
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Konflikt — zasób zmieniony w innym miejscu albo już istnieje
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [errorCode, message, timestamp, traceId]
      properties:
        errorCode: { type: string, examples: [TRACK_NOT_FOUND] }
        message: { type: string }
        timestamp: { type: string, format: date-time }
        traceId: { type: [string, 'null'] }
    TrackResponse:
      type: object
      required: [spotifyId, title, artist, album, year, durationMs, popularity, explicit, albumImageUrl,
                 isrc, genreFamily, style, bpm, bpmSource, danceability, musicalKey,
                 camelot, tempoClass, energy, lyricsTheme, descriptionPl, confidence,
                 enrichedAt, modelUsed, enrichVersion]
      properties:
        spotifyId: { type: string }
        title: { type: [string, 'null'] }
        artist: { type: [string, 'null'] }
        album: { type: [string, 'null'] }
        year: { type: [integer, 'null'], format: int32 }
        durationMs: { type: [integer, 'null'], format: int32 }
        popularity: { type: [integer, 'null'], format: int32 }
        explicit: { type: [boolean, 'null'] }
        albumImageUrl: { type: [string, 'null'] }
        isrc: { type: [string, 'null'] }
        genreFamily: { type: [string, 'null'] }
        style: { type: [string, 'null'] }
        bpm: { type: [integer, 'null'], format: int32 }
        bpmSource: { type: [string, 'null'] }
        danceability: { type: [number, 'null'] }
        musicalKey: { type: [string, 'null'] }
        camelot:
          type: [string, 'null']
          description: Liczona z `musicalKey` przez backend — nie kolumna w bazie
        tempoClass: { type: [string, 'null'] }
        energy: { type: [string, 'null'] }
        lyricsTheme: { type: [string, 'null'] }
        descriptionPl: { type: [string, 'null'] }
        confidence: { type: [string, 'null'] }
        enrichedAt: { type: [string, 'null'], format: date-time }
        modelUsed: { type: [string, 'null'] }
        enrichVersion: { type: [integer, 'null'], format: int32 }
...
    TrackMetricsResponse:
      type: object
      required: [spotifyId, bpm, musicalKey, camelot, danceability, energy, valence, acousticness,
                 instrumentalness, speechiness, liveness, loudnessDb, timeSignature,
                 source, importedAt]
      properties:
        spotifyId: { type: string }
        bpm: { type: [number, 'null'] }
        musicalKey: { type: [string, 'null'] }
        camelot: { type: [string, 'null'] }
        danceability: { type: [number, 'null'] }
        energy: { type: [number, 'null'] }
        valence: { type: [number, 'null'] }
        acousticness: { type: [number, 'null'] }
        instrumentalness: { type: [number, 'null'] }
        speechiness: { type: [number, 'null'] }
        liveness: { type: [number, 'null'] }
        loudnessDb: { type: [number, 'null'] }
        timeSignature: { type: [integer, 'null'], format: int32 }
        source: { type: [string, 'null'] }
        importedAt: { type: string, format: date-time }
...
    PlaylistSummaryResponse:
      type: object
      required: [id, name, spotifyPlaylistId, createdAt, trackCount, version]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        spotifyPlaylistId: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        trackCount: { type: integer, format: int64 }
        version: { type: integer, format: int32 }
    PlaylistTrackResponse:
      type: object
      required: [position, djSlot, djSlotOverride, metrics, track]
      properties:
        position: { type: integer, format: int32 }
        djSlot: { type: [string, 'null'] }
        djSlotOverride: { type: [string, 'null'] }
        metrics:
          oneOf:
            - { $ref: '#/components/schemas/TrackMetricsResponse' }
            - { type: 'null' }
        track: { $ref: '#/components/schemas/TrackResponse' }
    PlaylistResponse:
      type: object
      required: [id, name, spotifyPlaylistId, createdAt, version, tracks]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        spotifyPlaylistId: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        version: { type: integer, format: int32 }
        tracks:
          type: array
          items: { $ref: '#/components/schemas/PlaylistTrackResponse' }
    SavePlaylistRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        version:
          type: [integer, 'null']
          format: int32
          description: Wymagana przy zmianie nazwy; pomijana przy tworzeniu
    AddPlaylistTrackRequest:
      type: object
      required: [spotifyId]
      properties:
        spotifyId: { type: string, minLength: 1, maxLength: 64 }
    ReorderPlaylistRequest:
      type: object
      required: [spotifyIds, version]
      properties:
        spotifyIds:
          type: array
          minItems: 1
          items: { type: string, minLength: 1 }
        version: { type: integer, format: int32 }
    PlaylistExportResponse:
      type: object
      required: [playlistId, spotifyPlaylistId, name, exportedTracks, created, spotifyUrl]
      properties:
        playlistId: { type: integer, format: int64 }
        spotifyPlaylistId: { type: string }
        name: { type: string }
        exportedTracks: { type: integer, format: int32 }
        created:
          type: boolean
          description: true — playlista założona, false — nadpisana
        spotifyUrl: { type: string }
    SetFilters:
      type: object
      description: Filtry puli kandydatów, wspólne dla propozycji, uzupełniania i dobierania.
      properties:
        search: { type: [string, 'null'] }
        genreFamily: { type: [string, 'null'] }
        bpmMin: { type: [integer, 'null'], format: int32 }
        bpmMax: { type: [integer, 'null'], format: int32 }
        tempoClass: { type: [string, 'null'] }
        energy: { type: [string, 'null'] }
        inLibrary: { type: [boolean, 'null'] }
        ratingMin: { type: [integer, 'null'], format: int32, minimum: 1, maximum: 5 }
        tag: { type: [string, 'null'] }
        camelot: { type: [string, 'null'] }
        camelotCompatible: { type: [boolean, 'null'] }
        valenceMin: { type: [number, 'null'] }
        valenceMax: { type: [number, 'null'] }
        instrumentalMin: { type: [number, 'null'] }
        livenessMax: { type: [number, 'null'] }
    SetProposalRequest:
      allOf:
        - $ref: '#/components/schemas/SetFilters'
        - type: object
          required: [targetMinutes]
          properties:
            targetMinutes: { type: integer, format: int32, minimum: 15, maximum: 720 }
            curve: { type: [string, 'null'] }
            seed: { type: [integer, 'null'], format: int64 }
    SetFillRequest:
      allOf:
        - $ref: '#/components/schemas/SetFilters'
        - type: object
          required: [targetMinutes]
          properties:
            targetMinutes: { type: integer, format: int32, minimum: 15, maximum: 720 }
            curve: { type: [string, 'null'] }
            seed: { type: [integer, 'null'], format: int64 }
    SetSuggestionRequest:
      allOf:
        - $ref: '#/components/schemas/SetFilters'
        - type: object
          properties:
            position:
              type: [integer, 'null']
              format: int32
              minimum: 0
              description: Brak wartości znaczy „na koniec setu"
            limit: { type: [integer, 'null'], format: int32, minimum: 1, maximum: 20 }
    ProposedTrackResponse:
      type: object
      required: [position, djSlot, track]
      properties:
        position: { type: integer, format: int32 }
        djSlot: { type: [string, 'null'] }
        track: { $ref: '#/components/schemas/TrackResponse' }
    SetProposalResponse:
      type: object
      required: [trackCount, totalDurationMs, targetDurationMs, seed, notes, tracks]
      properties:
        trackCount: { type: integer, format: int32 }
        totalDurationMs: { type: integer, format: int64 }
        targetDurationMs: { type: integer, format: int64 }
        seed: { type: integer, format: int64 }
        notes: { type: array, items: { type: string } }
        tracks:
          type: array
          items: { $ref: '#/components/schemas/ProposedTrackResponse' }
    SetFillResponse:
      type: object
      required: [currentTrackCount, currentDurationMs, addedTrackCount, totalDurationMs,
                 targetDurationMs, seed, notes, tracks]
      properties:
        currentTrackCount: { type: integer, format: int32 }
        currentDurationMs: { type: integer, format: int64 }
        addedTrackCount: { type: integer, format: int32 }
        totalDurationMs: { type: integer, format: int64 }
        targetDurationMs: { type: integer, format: int64 }
        seed: { type: integer, format: int64 }
        notes: { type: array, items: { type: string } }
        tracks:
          type: array
          items: { $ref: '#/components/schemas/ProposedTrackResponse' }
    SuggestedTrackResponse:
      type: object
      required: [djSlot, bpmDelta, harmonic, track]
      properties:
        djSlot: { type: [string, 'null'] }
        bpmDelta: { type: [integer, 'null'], format: int32 }
        harmonic: { type: [boolean, 'null'] }
        track: { $ref: '#/components/schemas/TrackResponse' }
    SetSuggestionResponse:
      type: object
      required: [position, suggestions]
      properties:
        position: { type: integer, format: int32 }
        suggestions:
          type: array
          items: { $ref: '#/components/schemas/SuggestedTrackResponse' }
...